+++ packages/video-http-ingestion/src/VideoHttpIngestionAppliance.js
@@ -44,4 +44,4 @@ class VideoHttpIngestionAppliance extends AbstractVideoIngestionAppliance {
 	}
 }
 
-export default VideoHttpIngestionAppliance
+export { VideoHttpIngestionAppliance }