+++ packages/ccextractor/CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Changed from default exports to named exports.
+
+### Added
+- Now exports the `CCExtractorLine` object because why not.
 
 ## [0.1.0] - 2020-09-09
 ### Added